backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0

# LLM provider (install to use LLM, otherwise uses mock)
# litellm supports OpenAI, Anthropic, Cohere, Google, and many more
# Install: pip install litellm
# litellm>=1.0.0

# For Langfuse integration (optional - for LLM observability)
# langfuse>=2.0.0

